requirements.txt
langchain==0.1.20 
langchain-community==0.0.38 
langchain-core==0.1.53 
langchain-text-splitters==0.0.2 
#chromadb==0.4.24 
pypdf==4.2.0 
langchain-google-genai==0.0.6
sentence-transformers==2.6.1
transformers==4.38.2
numpy==1.26.4
cryptography==42.0.5
chromadb==0.4.24 
posthog==3.5.0
python-dotenv